""
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@560 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/d/binding.d
@@ -51,7 +51,7 @@ export parens(binaryleft:string,binaryright:string,prec:int,scope:int,strength:i
      );
 
 HadError := false;
-unknown := Position("--unknown position--",ushort(0),uchar(0),uchar(reloaded));
+unknown := Position("--unknown position--",ushort(0),ushort(0),uchar(reloaded));
 export treePosition(e:ParseTree):Position := (
      while true do (
 	  when e
